langchain-service/requirements.txt
python-dotenv
Flask
langchain
openai
wikipedia
pyowm